refactor(auth): refactor auth context to improve separation of concerns
