+++ src/app/api/auth/login/route.ts
@@ -62,7 +62,7 @@ export async function POST(request: Request) {
     }
 
     // Create a user object without the password
-    const { password: _, ...userWithoutPassword } = user
+    const { ...userWithoutPassword } = user
 
     // Generate a mock token (in a real app, use JWT or similar)
     const token = `mock-token-${Date.now()}-${Math.random().toString(36).substring(2, 15)}`
