+++ src/features/auth/api/index.ts
@@ -1,0 +1,3 @@
+import { loginApi, resetPasswordApi, validateTokenApi } from "@/lib/api/auth";
+
+export { loginApi, resetPasswordApi, validateTokenApi };
